html: move markup of AST out of main.cpp
